requirements.txt
pandas>=1.3.0
motor>=2.5.0
pymongo>=3.12.0
yfinance>=0.1.70
aiohttp>=3.8.1
python-dateutil>=2.8.2
argparse>=1.4.0
aiofiles>=0.8.0
certifi>=2023.7.22
openpyxl>=3.1.2
sqlite3>=3.0.0
python-dotenv>=1.0.0 